Fix IsInSameTeamWith when the players are equal. Inline GetCard in Card.

+++ src/AI/Belot.AI.SmartPlayer/GlobalCounters.cs
@@ -4,6 +4,6 @@
     {
         public const int CountersCount = 8;
 
-        public static int[] Counters { get; set; } = new int[CountersCount];
+        public static long[] Counters { get; set; } = new long[CountersCount];
     }
 }
